% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryModel.R
\name{summaryModel}
\alias{summaryModel}
\title{Prints summary information about nominal or GPCM fit to data}
\usage{
summaryModel(opleLMA)
}
\arguments{
\item{opleLMA}{Object with results from ple.lma for "nominal", "gpcm", or "rasch"}
}
\description{
The output from fitting a model contains lots of information as a
list.  This function take model fit object and returns a summary
of model fit that is printed in console.  Information includes
model specification, information about data, convergence information,
and global fit statistics. This function is designed for Nominal,
generalized partial credit, and Rasch models.
}
